Moves testing utils to support module
